refactor: Restructure Loadable component
Relocate LoadingTodos and LoadingLabels to the colocated directory,
including the associated loadingSkeleton. The restructured directories
ensure each Loadable component has its own folder. These changes enhance
the organization and management within the project.
