+++ components/label/loadingLabels/index.tsx
@@ -1,7 +1,7 @@
+import { LoadingState } from '@components/loadable/loadingStates';
 import { optionsLoadingLabels } from '@options/loadingState';
 import { SmoothTransition } from '@ui/transitions/smoothTransition';
 import { Fragment } from 'react';
-import { LoadingState } from '.';
 
 export const LoadingLabels = () => {
   return (
