+++ components/todos/loadingTodos/index.tsx
@@ -1,7 +1,7 @@
+import { LoadingState } from '@components/loadable/loadingStates';
 import { optionsLoadingTodos } from '@options/loadingState';
 import { SmoothTransition } from '@ui/transitions/smoothTransition';
 import { Fragment as LoadingStateFragment } from 'react';
-import { LoadingState } from '.';
 
 export const LoadingTodos = () => {
   return (
